# [vm] postpone creation of handles for show/hide lists