fixes: #60470

for some reason it fixes deallocation of empty show/hide lists

before (ends with out of memory):
<img width="1489" alt="image" src="https://github.com/user-attachments/assets/430bf42b-53c4-437e-8f25-f535749d9bd0" />


after:
<img width="1450" alt="image" src="https://github.com/user-attachments/assets/07e5ab5c-7c87-4ae4-8eca-a4ce3f198236" />

---

- [x] I’ve reviewed the contributor guide and applied the relevant portions to this PR.

<details>
  <summary>Contribution guidelines:</summary><br>

- See our [contributor guide](https://github.com/dart-lang/sdk/blob/main/CONTRIBUTING.md) for general expectations for PRs.
- Larger or significant changes should be discussed in an issue before creating a PR.
- Contributions to our repos should follow the [Dart style guide](https://dart.dev/guides/language/effective-dart) and use `dart format`.

Note that this repository uses Gerrit for code reviews. Your pull request will be automatically converted into a Gerrit CL and a link to the CL written into this PR. The review will happen on Gerrit but you can also push additional commits to this PR to update the code review.
</details>