@2ZeroSix 2025-04-04
execution time under profiler was consistently better with both patches applied, i'm running a benchmark to see if the difference is real (will be done in approx 3.5 hours)