@2ZeroSix 2025-04-04
> execution time under profiler was consistently better with both patches applied, i'm running a benchmark to see if the difference is real (will be done in approx 3.5 hours)

diff between https://dart-review.googlesource.com/c/sdk/+/420321 and both fixes is not convincing enough yet, but overall these fixes seems to reduce gen_snapshot time by ~7% for our app

Will check against a commit that can be built with a version without these fixes and try to remove additional noise by measuring only the LoadProgram step tomorrow


<details><summary>result</summary>
<p>

script:
https://gist.github.com/2ZeroSix/9f448bd51b596432c30c7bd0b22d9066

output:
```
================================================================================
BENCHMARK RESULTS
================================================================================

gen_snapshot_both:
  Individual runs (seconds):
    Run 1: 345.40s
    Run 2: 336.66s
    Run 3: 330.03s
    Run 4: 337.52s
    Run 5: 421.26s # excluded
    Run 6: 351.93s
    Run 7: 333.67s
    Run 8: 335.12s
    Run 9: 332.88s
    Run 10: 320.78s
  Average: 335.99s (fastest)
  Median:  335.12s
  Min:     320.78s
  Max:     351.93s
  Range:   31.15s

gen_snapshot_mraleph:
  Individual runs (seconds):
    Run 1: 339.49s
    Run 2: 341.86s
    Run 3: 336.59s
    Run 4: 342.88s
    Run 5: 339.70s
    Run 6: 344.09s
    Run 7: 339.10s
    Run 8: 349.48s
    Run 9: 963.80s # excluded
    Run 10: 341.20s
  Average: 341.59s (1.66% slower than gen_snapshot_both)
  Median:  341.20s
  Min:     336.59s
  Max:     349.48s
  Range:   12.89s

gen_snapshot_2zerosix:
  Individual runs (seconds):
    Run 1: 353.91s
    Run 2: 388.05s
    Run 3: 387.13s
    Run 4: 343.74s
    Run 5: 348.80s
    Run 6: 360.28s
    Run 7: 356.25s
    Run 8: 345.39s
    Run 9: 442.23s # excluded
    Run 10: 361.02s
  Average: 360.50s (7.29% slower than gen_snapshot_both)
  Median:  356.25s
  Min:     343.74s
  Max:     388.05s
  Range:   44.31s

================================================================================
```

</p>
</details>