@mraleph 2025-04-07
@2ZeroSix thank you for those extensive measurements. I think we are going to land both. Just waiting for @rmacnak-google to review them :)